Refactor InsertObjectMedia() integration tests.
This is partly to make the object integration test smaller (#1282) and
partly to make room for the new InsertObjectMedia() tests that will also
accept the object metadata values (#1311).
